[#106] rename: PlaceCard2 -> PlaceCard 파일명 수정
