+++ src/components/tourism/PlaceCard.tsx
